Added vagrant-common as a shared folder to the volume vm.
Added code to the Vagrant bash scripts to delete the volumes on the volume VM.  It kind of stinks now and makes the logs noisy though.

Added a set of standard SSH keys to the Vagrant directory, so the volume could easily allow itself to be logged into by the client.  I'm not sure if our test VMs ports are open to outsides, This could be a security risk, so we'll need to remember to look into that if we run on machines where our test VMs ports are open to outsiders.  This is now called from run_ci.sh.  Changed the VagrantUtils.sh script so we could use the ssh functions on the host or volume VM in order to do this.

Added a new bash script to run the direct volume integration tests.

Added some san properties to the nova.conf.template file.

Created a utility method to poll until some condition. Not sure if I like it, but I couldn't think of a way to simulate a do until loop that wouldn't duplicate code.

Cleaned up the direct.py test code a bit.

Added "ISCSITargetNotDiscoverable" exception, which gets thrown if somehow the volume exists in the DB, the iscsi target ID exists in the DB, but the later can't get found when "iscsiadm -m discovery" is called.

Looked through the nova.volume.driver code, realized that in some places commands that should only be executed on the nova-compute or nova-volume node get executed one after another in the same call. That doesn't work for our purposes so I had to change this code a bit. Now the part of the code in discover_volume that finds the iscsi log-in properties (and tries to create a volume as if its on the nova-volume node) is in its own overridable method.

Made VolumeClient inherit from nova.db.base.Base so the driver could have its database set.