I moved all of the volume creation code out of the Nove compute manager and into the subclassed Reddwarf one.  The _run_instance method now is in charge of handling the volume, the compute instance, and the monitoring the guest install process, and can abort at any of these stages. There are (mostly) test cases on everything but the compute instance failing to create.
I also refactored the code for the subclassed manager to avoid passing instance and context around everwhere.  Hopefully very soon when we write unit tests for this it will be easier to test, but my main motivation was to make it easier to call the error logging code. Now in addition to calling LOG.audit, I added a call to the nofitier whenever something goes bad during instance creation.

Added the ability to poll_until to not require a second function argument if the first one returns a boolean.
nova/network/manager.py

Added the ability to stop and restart the services that until now have started early on in our tests and run forever. This is necessary because the stuff I'm testing defaults to long time limits for production, and I needed to change the flags to make those timeouts shorter in testing. I changed the guest_initialize_failure test to use this new ability to restart the compute service with different flag values.

I also updated the instance initialization tests use Proboscis v 1.2. Because of this it no longer needs global variables and the various junk that lived in the "ContainerTestInfo" object now lives in the test classes themselves.

Thanks to that, the original test case for the timed-out guest instance install could be consolidated into one class. It was then easy to extract a base class from it, and create a new case for when a volume fails. To do this I also needed to extract the volume service into an object living in test_config which could be stopped before the test and started afterwards.