Removed the part of the Nova compute manager that waited for Volumes, as it had snuck back in during the rebase.
Moved around a test in containers.py since the failure case is more quickly caught in the WaitForGuestInstallationToFinish test than the VerifyGuestStarted one.

Changed how the database list metadata was serialized to avoid validating it a second time needlessly.
Changed ReddwarfInstanceInitializer to include the compute manager in the constructor since it was passed in to all three public methods. I don't prefer it this way but think it will sit better with the team.

Changed a while loop to read a bit nicer.